skyemu-mcp/requirements.txt
mcp==1.5.0
requests==2.31.0
Pillow==10.2.0
